refactor: Move INotificationService to Common/Interfaces
Moved `INotificationService.cs` from `Valora.Application.Services` to `Valora.Application.Common.Interfaces` to strictly align with clean architecture boundaries where abstractions are centrally located and separate from concrete implementations.

Updated all call sites including `NotificationEndpoints.cs` and `NotificationEventHandlers.cs` to use the correct namespace.
