+++ backend/Valora.Infrastructure/Services/NotificationEventHandlers.cs
@@ -1,4 +1,3 @@
-using Valora.Application.Services;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
